move modules to root of their folder
