+++ src/modules/cache/cache.module.ts
@@ -3,9 +3,9 @@ import { CacheModule } from '@nestjs/cache-manager'
 import { type RedisClientOptions } from 'redis'
 import { redisStore } from 'cache-manager-redis-store'
 import { ConfigModule, ConfigService } from '@nestjs/config'
-import { RoleRepository } from '../../modules/roles/repositories/role.repository.js'
-import { UserRepository } from '../../modules/users/repositories/user.repository.js'
-import { RedisCacheService } from './cache.js'
+import { RoleRepository } from '../roles/repositories/role.repository.js'
+import { UserRepository } from '../users/repositories/user.repository.js'
+import { RedisCacheService } from './services/cache.service.js'
 
 @Module({
   imports: [CacheModule.registerAsync<RedisClientOptions>({
@@ -19,7 +19,11 @@ import { RedisCacheService } from './cache.js'
     inject: [ConfigService]
   })],
   controllers: [],
-  providers: [RedisCacheService, RoleRepository, UserRepository],
+  providers: [
+    RedisCacheService,
+    RoleRepository,
+    UserRepository
+  ],
   exports: [RedisCacheService]
 })
 export class RedisCacheModule {}
