+++ src/modules/mail/mail.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common'
-import { MailService } from '../services/mail.service.js'
-import { ScalewayMailClient } from '../clients/scaleway-mail.client.js'
-import { MjmlRenderer } from '../renderer/mjml.renderer.js'
+import { ScalewayMailClient } from './clients/scaleway-mail.client.js'
+import { MjmlRenderer } from './renderer/mjml.renderer.js'
+import { MailService } from './services/mail.service.js'
 
 @Module({
   providers: [
