+++ src/modules/roles/role.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common'
 import { TypeOrmModule } from '@nestjs/typeorm'
 import { UserRepository } from '../users/repositories/user.repository.js'
-import { RedisCacheModule } from '../../utils/cache/cache.module.js'
+import { RedisCacheModule } from '../cache/cache.module.js'
 import { Role } from './entities/role.entity.js'
 import { RoleController } from './controllers/role.controller.js'
 import { RoleService } from './services/role.service.js'
